+++ src/main/java/fr/alasdiablo/diolib/event/IEvent.java
@@ -1,0 +1,7 @@
+package fr.alasdiablo.diolib.event;
+
+import net.minecraftforge.eventbus.api.Event;
+
+public interface IEvent {
+    void init(Event e);
+}